refac: remove redundancy form the public API
There was no need for static methods in the Stream object doing the same
functions as an instance of Stream. The API is now more streamlined,
with the only static method being Stream.fromElements.

Also, the different kinds of streams is now implemented through
subclasses of Stream, instead of through several generator functions.