fix: fix reduce and add index to find and forEach
This commit fixes the reduce functions of Stream and AsyncStream to
deal with corner cases in the same way as Array.prototype.reduce. That
is, when the backing iterable is empty, the reduce method returns the
initial value, if it exists, or throws an exception if it doesn't.

Also, this commit modifies the find and forEach methods of Stream and
AsyncStream to pass a second "index" parameter to the function they
receive as a parameter. This makes the find and forEach methods conform
to the JavaScript standard for such functions.